+++ test/queries/MetricsQueries.test.ts
@@ -1,20 +1,23 @@
 import 'reflect-metadata';
 import { expect } from 'chai';
+import addDays from 'date-fns/addDays';
 
-import { MetricsRepository } from '../../src/repositories/MetricsRepository';
+import { MetricsQueries } from '../../src/queries/MetricsQueries';
 import { getTestContainer } from '../helpers/getTestContainer';
 import { setupDatabase, teardownDatabase } from '../helpers/databaseHelpers';
 import Block from '../../src/models/Block';
 import { inputForBlockModel } from '../fixtures/inputForBlockModel';
-import addDays from 'date-fns/addDays';
+import leaves from '../mocks/leaves.json';
+import Leaf from '../../src/models/Leaf';
 
 const container = getTestContainer();
-container.bind(MetricsRepository).toSelf();
-const metricsRepository = container.get(MetricsRepository);
+container.bind(MetricsQueries).toSelf();
+const metricsQueries = container.get(MetricsQueries);
 
-describe('MetricsRepository', () => {
+describe('MetricsQueries', () => {
   before(async () => {
     await setupDatabase();
+    await Leaf.deleteMany();
 
     await Block.create([
       {
@@ -30,17 +33,20 @@ describe('MetricsRepository', () => {
     await Block.create([
       { ...inputForBlockModel, _id: 'block::4', blockId: 4, voters: ['0xA405324F4b6EB7Bc76f1964489b3769cfc71445H'] },
     ]);
+
+    await Promise.all(leaves.map((leaf) => Leaf.create(leaf)));
   });
 
   after(async () => {
     await Block.deleteMany({ _id: ['block::1', 'block::2', 'block::3', 'block::4'] });
+    await Leaf.deleteMany();
     await teardownDatabase();
   });
 
   describe('#getVotersCount', () => {
     it('returns only the voter count from finalized blocks', async () => {
       const endDateFormat = addDays(inputForBlockModel.dataTimestamp, 1);
-      const votersCount = await metricsRepository.getVotersCount({
+      const votersCount = await metricsQueries.getVotersCount({
         startDate: inputForBlockModel.dataTimestamp,
         endDate: endDateFormat,
       });
@@ -53,6 +59,20 @@ describe('MetricsRepository', () => {
         },
         { _id: '0xA405324F4b6EB7Bc76f1964489b3769cfc71445F', count: 2 },
       ]);
+    });
+  });
+
+  describe('#getKeysCount', () => {
+    it('returns the keys count from leafs with blockId bigger than given', async () => {
+      const startBlockNumber = 169269;
+      const endBlockNumber = 169271;
+      const keysCount = await metricsQueries.getKeysCount(startBlockNumber, endBlockNumber);
+
+      expect(keysCount).to.be.an('array');
+      expect(keysCount).to.deep.include({
+        _id: 'BCH-BTC',
+        count: 2,
+      });
     });
   });
 });