+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Create metrics/keys-frequency endpoint to get frequency of appearance of keys in the latest 100 blocks
 
 ## [4.17.1] - 2022-05-04
 ### Changed